Skip GFF3 generation errors

+++ modules/LRG/Pipeline/MoveXMLFiles.pm
@@ -79,7 +79,9 @@ sub run {
       my $fasta_dir = ($type_dir eq 'stalled') ? "$ftp_subdir/fasta/" : "$ftp_dir/fasta/";
       $self->run_cmd("cp $new_xml_dir/fasta/$lrg_id.fasta $fasta_dir");
       # Copy GFF to FTP site
-      $self->run_cmd("cp $new_xml_dir/gff/$lrg_id\_*.gff $ftp_dir/.ensembl_internal/");
+      if (glob("$new_xml_dir/gff/$lrg_id\_*.gff")) { # Check if at least one file exists
+        $self->run_cmd("cp $new_xml_dir/gff/$lrg_id\_*.gff $ftp_dir/.ensembl_internal/");
+      }
     }
   }
 }
